// src/app/models/issue.model.ts
export interface Issue {
  userId: string;
  userName: string;
  circle: string;
  ba: string;
  phone: string;
  numOfUsersAffected: number;
  issueCategories: string;
  issueDate: Date;
  resolutionTime: string;
  issueDescription: string;
  status: string;
  remarks: string;
  ticketCreatedBy: string;
}